Refactor: 포스트/댓글 삭제 확인 모달 중복 제거
